requirements.txt
coverage
Django
django-cors-headers
djangorestframework
drf-yasg
psycopg2-binary
python-dotenv
django-filter
flake8